TakeDamage(float)

+++ Gyeongseong97/Enemy.h
@@ -19,7 +19,7 @@ public:
 	Enemy();
 	Enemy(int x, int y, int health, float speed, int killScore);
 
-	virtual void TakeDamage(int damage);
+	virtual void TakeDamage(float damage);
 	void Update() override;
 	void OnCollision(GameObject& other) override;
 	void Destroy() override;
